+++ packages/picasso-charts/src/BarChart/test.tsx
@@ -9,9 +9,24 @@ const TEST_DATA = [
 describe('BarChart', () => {
   it('formats data correctly', () => {
     const expectedData = [
-      { name: 'Moscow', infected: 1, recovered: 2 },
-      { name: 'Berlin', infected: 2, recovered: 5 },
-      { name: 'London', infected: 7, recovered: 8 }
+      {
+        name: 'Moscow',
+        infected: 1,
+        recovered: 2,
+        value: { infected: 1, recovered: 2 }
+      },
+      {
+        name: 'Berlin',
+        infected: 2,
+        recovered: 5,
+        value: { infected: 2, recovered: 5 }
+      },
+      {
+        name: 'London',
+        infected: 7,
+        recovered: 8,
+        value: { infected: 7, recovered: 8 }
+      }
     ]
 
     expect(formatData<'infected' | 'recovered'>(TEST_DATA)).toEqual(
