+++ packages/topkit-analytics-charts/src/CategoriesChartTooltip/index.ts
@@ -1,0 +1,1 @@
+export { default } from './CategoriesChartTooltip'
